edits

+++ layer3/layer3_gan_ref_back.py
@@ -309,7 +309,7 @@ gan_model = define_gan(g_model, d_model, image_shape)
 #predict results
 model = load_model('../../models_3/model_ref_back.h5')
 # print(dataset[0][0].shape)
-img = Image.fromarray(dataset[0][0])
+img = Image.fromarray(dataset[0][0].astype(np.uint8))
 result = model.predict(dataset[0][0])
 # result = Image.fromarray(result)
 # result.save('prediction.png')
